Move Connect back to comm module
